Logic improvement
Hi @shaileshcheke. Nice project. Still, there is no need for else statement. 
Think of it like this.

If the condition is true then will return y and jump out of the function so the next line won't be executed.
Else if the if tends to be false then return x.
I hope I made it clear.